+++ architecture/src/Sienar.Architecture.EntityFramework/Data/EfEntityReadAllActor.cs
@@ -6,7 +6,6 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Sienar.Hooks;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Data;
