+++ architecture/src/Sienar.Architecture.EntityFramework/Data/IEfFilterProcessor.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Linq;
 using System.Linq.Expressions;
-using Sienar.Data;
 using Sienar.Hooks;
 
-namespace Sienar.Processors;
+namespace Sienar.Data;
 
 /// <summary>
 /// A processor which converts a <see cref="Filter"/> into EF queries for the given <c>TEntity</c>
