+++ plugins/identity/src/Sienar.Plugins.Identity.Client/Identity/Processors/ClientAccountLockoutProcessor.cs
@@ -1,8 +1,8 @@
 ﻿using System.Threading.Tasks;
+using Sienar.Data;
 using Sienar.Identity.Requests;
 using Sienar.Identity.Results;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
