+++ plugins/identity/src/Sienar.Plugins.Identity.Client/Identity/Processors/ClientInitiateEmailChangeProcessor.cs
@@ -1,9 +1,9 @@
 ﻿#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
 using System.Threading.Tasks;
+using Sienar.Data;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
