+++ plugins/identity/src/Sienar.Plugins.Identity.Client/Identity/Processors/ClientLogoutProcessor.cs
@@ -1,10 +1,10 @@
 ﻿#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
 using System.Threading.Tasks;
+using Sienar.Data;
 using Sienar.Extensions;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
