+++ plugins/identity/src/Sienar.Plugins.Identity.Client/Identity/Processors/LoadUserDataProcessor.cs
@@ -2,10 +2,10 @@
 
 using System.Linq;
 using System.Threading.Tasks;
+using Sienar.Data;
 using Sienar.Hooks;
 using Sienar.Identity.Results;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 using Sienar.Security;
 using Sienar.Ui;
