+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/ForgotPasswordProcessor.cs
@@ -10,7 +10,6 @@ using Sienar.Errors;
 using Sienar.Extensions;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
