+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/GetAccountDataProcessor.cs
@@ -3,9 +3,9 @@
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using Sienar.Data;
 using Sienar.Identity.Results;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
