+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/GetLockoutReasonsProcessor.cs
@@ -8,7 +8,6 @@ using Sienar.Errors;
 using Sienar.Identity.Requests;
 using Sienar.Identity.Results;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
