+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/LockUserAccountProcessor.cs
@@ -9,7 +9,6 @@ using Sienar.Errors;
 using Sienar.Identity.Requests;
 using Sienar.Email;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
