+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/LockoutReasonFilterProcessor.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Linq.Expressions;
 using Sienar.Data;
 using Sienar.Extensions;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
