+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/LoginProcessor.cs
@@ -12,7 +12,6 @@ using Sienar.Extensions;
 using Sienar.Identity.Requests;
 using Sienar.Identity.Results;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
