+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/ManuallyConfirmUserAccountProcessor.cs
@@ -5,7 +5,6 @@ using Sienar.Data;
 using Sienar.Errors;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
