+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/PersonalDataProcessor.cs
@@ -10,7 +10,6 @@ using Sienar.Errors;
 using Sienar.Identity.Hooks;
 using Sienar.Identity.Results;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
