+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/RegisterProcessor.cs
@@ -10,7 +10,6 @@ using Sienar.Email;
 using Sienar.Extensions;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
