+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/ResetPasswordProcessor.cs
@@ -8,7 +8,6 @@ using Sienar.Email;
 using Sienar.Errors;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
