+++ plugins/identity/src/Sienar.Plugins.Identity.Server/Identity/Processors/SienarRoleFilterProcessor.cs
@@ -4,7 +4,6 @@ using System;
 using System.Linq;
 using System.Linq.Expressions;
 using Sienar.Data;
-using Sienar.Processors;
 
 namespace Sienar.Identity.Processors;
